+++ src/Messaging.Abstractions/IEnvelopePublicationPolicy.cs
@@ -1,0 +1,8 @@
+namespace ArchiScrapper.Messaging.Abstractions;
+
+public interface IEnvelopePublicationPolicy<TPayload>
+{
+    IEnvelopePublicationErrorHandler<TPayload> ErrorHandler { get; }
+
+    IEnvelopePublicationTelemetry<TPayload>? Telemetry { get; }
+}