+++ src/Messaging.Abstractions/IEnvelopePublicationPolicyBuilder.cs
@@ -1,0 +1,10 @@
+namespace ArchiScrapper.Messaging.Abstractions;
+
+public interface IEnvelopePublicationPolicyBuilder<TPayload>
+{
+    IEnvelopePublicationPolicyBuilder<TPayload> UseErrorHandler(IEnvelopePublicationErrorHandler<TPayload> errorHandler);
+
+    IEnvelopePublicationPolicyBuilder<TPayload> UseTelemetry(IEnvelopePublicationTelemetry<TPayload> telemetry);
+
+    IEnvelopePublicationPolicy<TPayload> Build();
+}