+++ src/Messaging.Abstractions/IEnvelopePublisher.cs
@@ -9,14 +9,30 @@ public interface IEnvelopePublisher<TPayload>
         Func<TPayload, string> payloadSerializer,
         IEnvelopePublicationErrorHandler<TPayload>? errorHandler = null);
 
+    RawEnvelope PublishInlineWithPolicy(
+        TypedEnvelope<TPayload> source,
+        Func<TPayload, string> payloadSerializer,
+        IEnvelopePublicationPolicy<TPayload> policy);
+
     RawEnvelope Publish(
         TypedEnvelope<TPayload> source,
         Func<TPayload, string> payloadSerializer,
         IEnvelopePublicationErrorHandler<TPayload>? errorHandler = null);
+
+    RawEnvelope PublishWithPolicy(
+        TypedEnvelope<TPayload> source,
+        Func<TPayload, string> payloadSerializer,
+        IEnvelopePublicationPolicy<TPayload> policy);
 
     RawEnvelope PublishWithReference(
         TypedEnvelope<TPayload> source,
         Func<TPayload, string> payloadSerializer,
         string payloadReference,
         IEnvelopePublicationErrorHandler<TPayload>? errorHandler = null);
+
+    RawEnvelope PublishWithReferenceWithPolicy(
+        TypedEnvelope<TPayload> source,
+        Func<TPayload, string> payloadSerializer,
+        string payloadReference,
+        IEnvelopePublicationPolicy<TPayload> policy);
 }